:bug: Python 3 compatibility in effects

+++ octoprint_ws281x_led_status/effects/progress.py
@@ -1,5 +1,5 @@
 # Print and heat up progress?
-from __future__ import absolute_import, unicode_literals
+from __future__ import absolute_import, unicode_literals, division
 import time
 
 from octoprint_ws281x_led_status.util import blend_two_colors
@@ -8,8 +8,8 @@ from octoprint_ws281x_led_status.util import blend_two_colors
 def progress(strip, queue, value, progress_color, base_color, max_brightness=255):
     strip.setBrightness(max_brightness)
     num_pixels = strip.numPixels()
-    upper_bar = round((value / 100) * num_pixels)
-    lower_base = round(((100 - value) / 100) * num_pixels)
+    upper_bar = int(round((value / 100) * num_pixels))
+    lower_base = int(round(((100 - value) / 100) * num_pixels))
     if upper_bar + lower_base != strip.numPixels():
         print("Progress sanity check failed!, (bar){} +  (base){} = (total){} != (strip){}".format(
             upper_bar, lower_base, upper_bar + lower_base, strip.numPixels()))
